Implemented ModelNotFoundException.

+++ project/src/net/unikit/database/implementations/AbstractModelManagerImpl.java
@@ -1,6 +1,7 @@
 package net.unikit.database.implementations;
 
 import com.google.common.collect.ImmutableList;
+import net.unikit.database.exceptions.ModelNotFoundException;
 import net.unikit.database.interfaces.entities.AbstractModel;
 import net.unikit.database.interfaces.managers.AbstractModelManager;
 import org.hibernate.HibernateException;
@@ -12,7 +13,7 @@ import java.io.Serializable;
 import java.util.List;
 
 public abstract class AbstractModelManagerImpl<EntityType extends AbstractModel, IdType,
-        BaseEntityType, BaseIdType extends Serializable> implements AbstractModelManager<EntityType, IdType> {
+        BaseEntityType extends AbstractModel, BaseIdType extends Serializable> implements AbstractModelManager<EntityType, IdType> {
     private SessionFactory sessionFactory;
 
     protected AbstractModelManagerImpl(SessionFactory sessionFactory) {
@@ -20,10 +21,10 @@ public abstract class AbstractModelManagerImpl<EntityType extends AbstractModel,
     }
 
     interface TransactionAction<ResultType> {
-        ResultType run(Session session);
+        ResultType run(Session session) throws ModelNotFoundException;
     }
 
-    private <ResultType> ResultType doTransaction(TransactionAction<ResultType> transactionAction) {
+    private <ResultType> ResultType doTransaction(TransactionAction<ResultType> transactionAction) throws ModelNotFoundException {
         Session session = sessionFactory.openSession();
         Transaction transaction = null;
         ResultType result = null;
@@ -35,7 +36,7 @@ public abstract class AbstractModelManagerImpl<EntityType extends AbstractModel,
         } catch (HibernateException e) {
             if (transaction != null)
                 transaction.rollback();
-            e.printStackTrace();
+            throw e;
         } finally {
             if (session != null)
                 session.close();
@@ -46,34 +47,44 @@ public abstract class AbstractModelManagerImpl<EntityType extends AbstractModel,
 
     @Override
     public List<EntityType> getAllEntities() {
-        return doTransaction(new TransactionAction<List<EntityType>>() {
-            @Override
-            public List<EntityType> run(Session session) {
-                List<EntityType> entities = session.createQuery("FROM " + getAnnotatedClass().getSimpleName()).list();
-                return ImmutableList.copyOf(entities);
-            }
-        });
+        try {
+            return doTransaction(new TransactionAction<List<EntityType>>() {
+                @Override
+                public List<EntityType> run(Session session) {
+                    List<EntityType> entities = session.createQuery("FROM " + getAnnotatedClass().getSimpleName()).list();
+                    return ImmutableList.copyOf(entities);
+                }
+            });
+        } catch (ModelNotFoundException e) {
+            // NOTE: This exception will never been thrown!
+            e.printStackTrace();
+            return null;
+        }
     }
 
     @Override
-    public EntityType getEntity(IdType id) {
+    public EntityType getEntity(IdType id) throws ModelNotFoundException {
         return doTransaction(new TransactionAction<EntityType>() {
             @Override
-            public EntityType run(Session session) {
+            public EntityType run(Session session) throws ModelNotFoundException {
                 BaseIdType baseId = createBaseIdFromIdType(id);
                 EntityType entity = (EntityType) session.get(getAnnotatedClass(), baseId);
+                if (entity == null)
+                    throw new ModelNotFoundException(entity);
                 return entity;
             }
         });
     }
 
     @Override
-    public void updateEntity(EntityType entity) {
+    public void updateEntity(EntityType entity) throws ModelNotFoundException {
         doTransaction(new TransactionAction<Void>() {
             @Override
-            public Void run(Session session) {
+            public Void run(Session session) throws ModelNotFoundException {
                 BaseIdType baseId = createBaseIdFromIdType((IdType) entity.getId());
                 BaseEntityType entityOld = (BaseEntityType) session.get(getAnnotatedClass(), baseId);
+                if (entityOld == null)
+                    throw new ModelNotFoundException(entityOld);
                 updateDatabaseFields(entityOld, (BaseEntityType) entity);
                 session.update(entityOld);
                 return null;
@@ -82,12 +93,14 @@ public abstract class AbstractModelManagerImpl<EntityType extends AbstractModel,
     }
 
     @Override
-    public void deleteEntity(EntityType entity) {
+    public void deleteEntity(EntityType entity) throws ModelNotFoundException {
         doTransaction(new TransactionAction<Void>() {
             @Override
-            public Void run(Session session) {
+            public Void run(Session session) throws ModelNotFoundException {
                 BaseIdType baseId = createBaseIdFromIdType((IdType) entity.getId());
                 EntityType entityOld = (EntityType) session.get(getAnnotatedClass(), baseId);
+                if (entityOld == null)
+                    throw new ModelNotFoundException(entityOld);
                 session.delete(entityOld);
                 return null;
             }
@@ -96,13 +109,19 @@ public abstract class AbstractModelManagerImpl<EntityType extends AbstractModel,
 
     @Override
     public IdType addEntity(EntityType entity) {
-        return doTransaction(new TransactionAction<IdType>() {
-            @Override
-            public IdType run(Session session) {
-                BaseIdType id = (BaseIdType) session.save(entity);
-                return createIdFromBaseIdType(id);
-            }
-        });
+        try {
+            return doTransaction(new TransactionAction<IdType>() {
+                @Override
+                public IdType run(Session session) {
+                    BaseIdType id = (BaseIdType) session.save(entity);
+                    return createIdFromBaseIdType(id);
+                }
+            });
+        } catch (ModelNotFoundException e) {
+            // NOTE: This exception will never been thrown!
+            e.printStackTrace();
+            return null;
+        }
     }
 
     public abstract EntityType createEntity();
